+++ ref/randombytes.h
@@ -1,6 +1,6 @@
 #ifndef SPX_RANDOMBYTES_H
 #define SPX_RANDOMBYTES_H
 
-extern void randombytes(unsigned char * x,unsigned long long xlen);
+extern void randombytes(unsigned char * x, size_t xlen);
 
 #endif
